+++ server/internal/websocket/session.go
@@ -12,6 +12,18 @@ func (h *MessageHandler) SessionCreated(id string, session types.Session) error 
 		return err
 	}
 
+	// notify all about what is locked
+	for resource, id := range h.locked {
+		if err := session.Send(message.AdminLock{
+			Event:    event.ADMIN_LOCK,
+			ID:       id,
+			Resource: resource,
+		}); err != nil {
+			h.logger.Warn().Str("id", id).Err(err).Msgf("sending event %s has failed", event.ADMIN_LOCK)
+			return err
+		}
+	}
+
 	if session.Admin() {
 		// send screen configurations if admin
 		if err := h.screenConfigurations(id, session); err != nil {
@@ -21,17 +33,6 @@ func (h *MessageHandler) SessionCreated(id string, session types.Session) error 
 		// send broadcast status if admin
 		if err := h.boradcastStatus(session); err != nil {
 			return err
-		}
-
-		// if locked, notify admin about that
-		if h.locked {
-			if err := session.Send(message.Admin{
-				Event: event.ADMIN_LOCK,
-				ID:    id,
-			}); err != nil {
-				h.logger.Warn().Str("id", id).Err(err).Msgf("sending event %s has failed", event.ADMIN_LOCK)
-				return err
-			}
 		}
 	}